short -> Integer

+++ backend/src/main/java/hu/uni/eku/tzs/controller/dto/BillDto.java
@@ -14,6 +14,6 @@ public class BillDto {
     private LocalDate leave;
     private String firstName;
     private String surName;
-    private short numberOfDays;
+    private Integer numberOfDays;
     private Integer totalAmount;
 }
