+++ backend/src/main/java/hu/uni/eku/tzs/controller/dto/BillRecordRequestDto.java
@@ -18,6 +18,6 @@ public class BillRecordRequestDto {
     private LocalDate leave;
     private String firstName;
     private String surName;
-    private short numberOfDays;
+    private Integer numberOfDays;
     private Integer totalAmount;
 }
